update auto cache logic

+++ store/insert.ts
@@ -19,8 +19,6 @@ import {
   isRadiant,
   getPatchIndex,
   redisCount,
-  redisCountDistinct,
-  isAutoCachePlayer,
   transformMatch,
 } from '../util/utility';
 import { getMatchRankTier } from './queries';
@@ -502,17 +500,18 @@ export async function insertMatch(
         match.players.map(async (p) => {
           const account_id = pgroup[p.player_slot]?.account_id ?? p.account_id;
           if (account_id) {
-            await cassandra.execute(
-              `DELETE FROM player_temp WHERE account_id = ?`,
-              [account_id],
-              { prepare: true },
-            );
-            // Auto-cache players
-            if (await isAutoCachePlayer(redis, account_id)) {
+            const existsCheck = await cassandra.execute('SELECT account_id from player_temp WHERE account_id = ?', [account_id], { prepare: true });
+            // Auto-cache to update with this match if cache was already present
+            if (existsCheck.rows.length) {
               await addJob({
                 name: 'cacheQueue',
                 data: account_id.toString(),
               });
+              await cassandra.execute(
+                `DELETE FROM player_temp WHERE account_id = ?`,
+                [account_id],
+                { prepare: true },
+              );
             }
           }
         }),
